redesign dependencies installation / removing and bufixed backup

+++ remove.sh
@@ -1,32 +1,38 @@
 #!/bin/bash
 source $1
 
-# Check if the script is run as root
-
-#clear crontab
-sudo -u www-data crontab -r
-crontab -r
-
-rm -R /tmp/latest*
 rm -R $TEMP_DIR
-rm -R $NC_DATA_DIR
-rm -R $NC_DIR
-rm -R $BACKUP_ROOTDIR
-rm -R $BACKUP_DIR/daten
-systemctl stop apache2.service
-rm -R /etc/apache2
 
-docker kill onlyoffice-document-server
-docker container prune <<EOF
+if [ "$INSTALL_ONLYOFFICE" = "on" ]; then
+    docker kill onlyoffice-document-server
+    docker container prune <<EOF
 y
 EOF
-sleep 2
+    sleep 2
+fi
+ 
+if [ "$INSTALL_NC" = "on" ]; then
+    sudo -u www-data crontab -r
+    rm -R /tmp/latest*
+    rm -R $NC_DATA_DIR
+    rm -R $NC_DIR
+    systemctl stop apache2.service
+    rm -R /etc/apache2
 
-echo "Removing installed packages....."
-apt purge -y apache* php* borgbackup certbot python3-certbot-apache ffmpeg > /dev/null 2>&1
-apt purge -y maria*
+    echo "Removing NC-installed packages....."
+    apt purge -y apache* php* certbot python3-certbot-apache ffmpeg > /dev/null 2>&1
+    apt purge -y maria*
 
-echo "Removing done!"
+    echo "Removing NC-installation done!"
+fi
 
+if [ "$INSTALL_BACKUP" = "on" ]; then
+    crontab -r
+    echo "Removing Backup-Installation..."
+    rm -R $BACKUP_ROOTDIR
+    rm -R $BACKUP_DIR/daten
+    apt purge -y borgbackup expect > /dev/null 2>&1
+    echo "Removing Backup-Installation done!"
+fi
 
 exit 0
